requirements.txt
dnspython
python-dateutil
python-whois
six
dotenv
requests